sharePoint_upload.sh: remove 'Authorization: Bearer ...' header
This header is broken for two reasons: firstly it attempts to concatenate the
stoken variable into the header (which is empty), and secondly the presence of
this line causes Sharepoint Online to return a "Unsupported security token"
error. It is likely this is because the SP Online logic has changed such that
the token in the 'Authorization: Bearer ...' always takes priority over the
authentication cookies if both are present.